Fixed file uploading by adding an Upload folder, so all the uploaded files go straight there, also added a new controller called FileManagement, whose purpose is to do all actions associated with file management
