+++ Controllers/FileManagementController.cs
@@ -1,0 +1,48 @@
+﻿using Microsoft.AspNetCore.Mvc;
+
+namespace Student_Planner.Controllers
+{
+    public class FileManagementController : Controller
+    {
+        public IActionResult FileUpload()
+        {
+            return View();
+        }
+        [HttpPost]
+        public async Task<ActionResult> FileUpload(IFormFile file)
+        {
+            await UploadFile(file);
+            TempData["msg"] = "File Uploaded successfully.";
+            return View();
+        }
+        // Upload file on server
+        public async Task<bool> UploadFile(IFormFile file)
+        {
+            string path = "";
+            bool iscopied = false;
+            try
+            {
+                if (file != null && file.Length > 0)
+                {
+                    string filename = Guid.NewGuid() + Path.GetExtension(file.FileName);
+                    path = Path.GetFullPath(Path.Combine(Directory.GetCurrentDirectory(), "Upload"));
+                    using (var filestream = new FileStream(Path.Combine(path, filename), FileMode.Create))
+                    {
+                        await file.CopyToAsync(filestream);
+                    }
+                    iscopied = true;
+                }
+                else
+                {
+                    iscopied = false;
+                }
+            }
+            catch (Exception)
+            {
+                throw;
+            }
+            return iscopied;
+        }
+    }
+}
+
