+++ Controllers/HomeController.cs
@@ -35,44 +35,5 @@ namespace Student_Planner.Controllers
             return View(new ErrorViewModel { RequestId = Activity.Current?.Id ?? HttpContext.TraceIdentifier });
         }
 
-        public IActionResult FileUpload()
-        {
-            return View();
-        }
-        [HttpPost]
-        public async Task<ActionResult> FileUpload(IFormFile file)
-        {
-            await UploadFile(file);
-            TempData["msg"] = "File Uploaded successfully.";
-            return View();
-        }
-        // Upload file on server
-        public async Task<bool> UploadFile(IFormFile file)
-        {
-            string path = "";
-            bool iscopied = false;
-            try
-            {
-                if (file.Length > 0)
-                {
-                    string filename = Guid.NewGuid() + Path.GetExtension(file.FileName);
-                    path = Path.GetFullPath(Path.Combine(Directory.GetCurrentDirectory(), "Upload"));
-                    using (var filestream = new FileStream(Path.Combine(path, filename), FileMode.Create))
-                    {
-                        await file.CopyToAsync(filestream);
-                    }
-                    iscopied = true;
-                }
-                else
-                {
-                    iscopied = false;
-                }
-            }
-            catch (Exception)
-            {
-                throw;
-            }
-            return iscopied;
-        }
     }
 }
