+++ Program.cs
@@ -16,12 +16,16 @@ if (!app.Environment.IsDevelopment())
 app.UseHttpsRedirection();
 app.UseStaticFiles();
 
-app.UseRouting();
+app.UseRouting(
+    );
 
 app.UseAuthorization();
 
 app.MapControllerRoute(
     name: "default",
     pattern: "{controller=Home}/{action=Index}/{id?}");
+app.MapControllerRoute(
+    name: "fileUpload",
+    pattern: "{controller=FileManagement}/{action=FileUpload}/{id?}");
 
 app.Run();
