+++ Student Planner.csproj
@@ -8,6 +8,7 @@
   </PropertyGroup>
 
   <ItemGroup>
+    <Folder Include="Upload\" />
     <Folder Include="ReadFiles\" />
   </ItemGroup>
