+++ Views/FileManagement/FileUpload.cshtml
@@ -2,7 +2,7 @@
     ViewData["Title"] = "File upload";
 }
 
-<form method="post" enctype="multipart/form-data" asp-controller="Home" asp-action="FileUpload">
+<form method="post" enctype="multipart/form-data" asp-controller="FileManagement" asp-action="FileUpload">
     <div class="form-group">
         <div class="col-md-10">
             <p>Upload File</p>
