+++ Views/FileManagement/FileUpload.cshtml.cs
@@ -1,7 +1,7 @@
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.RazorPages;
 
-namespace Student_Planner.Views.Home
+namespace Student_Planner.Views.FileManagement
 {
     public class FileUploadModel : PageModel
     {
