+++ Views/Shared/_Layout.cshtml
@@ -26,7 +26,7 @@
                             <a class="nav-link text-dark" asp-area="" asp-controller="Home" asp-action="Privacy">Privacy</a>
                         </li>
                         <li class="nav-item">
-                            <a class="nav-link text-dark" asp-area="" asp-controller="Home" asp-action="FileUpload">File Upload</a>
+                            <a class="nav-link text-dark" asp-area="" asp-controller="FileManagement" asp-action="FileUpload">File Upload</a>
                         </li>
                     </ul>
                 </div>
